Fixes to move invocation of auditors to the core package.
