+++ core/src/main/java/gyro/core/GyroCore.java
@@ -22,7 +22,6 @@ import java.nio.file.Paths;
 
 import com.psddev.dari.util.Lazy;
 import com.psddev.dari.util.ThreadLocalStack;
-import gyro.core.command.GyroCommand;
 import org.apache.commons.lang3.StringUtils;
 
 public class GyroCore {
@@ -30,8 +29,6 @@ public class GyroCore {
     public static final String INIT_FILE = ".gyro/init.gyro";
 
     private static final ThreadLocalStack<GyroUI> UI = new ThreadLocalStack<>();
-
-    private static GyroCommand command;
 
     private static final Lazy<Path> HOME_DIRECTORY = new Lazy<Path>() {
 
@@ -71,14 +68,6 @@ public class GyroCore {
 
     public static GyroUI popUi() {
         return UI.pop();
-    }
-
-    public static GyroCommand getCommand() {
-        return command;
-    }
-
-    public static void setCommand(GyroCommand command) {
-        GyroCore.command = command;
     }
 
     public static Path getHomeDirectory() {
